opus-anima/raspberry/requirements.txt
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
dnspython==1.16.0
eventlet==0.26.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SocketIO==4.3.1
greenlet==0.4.16
idna==2.10
ipython==7.17.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
MouseInfo==0.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
prompt-toolkit==3.0.6
ptyprocess==0.6.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
Pygments==2.6.1
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pyserial==3.4
python-engineio==3.13.1
python-socketio==4.6.0
python3-xlib==0.15
PyTweening==1.0.3
requests==2.24.0
RPi.GPIO==0.7.0
six==1.15.0
SQLAlchemy==1.3.18
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
